Divide up search page into smaller components
