+++ frontend/react/src/index.tsx
@@ -1,6 +1,5 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import { BrowserRouter } from "react-router-dom";
 import App from "./App";
 
 ReactDOM.render(<App />, document.getElementById("root"));
