+++ frontend/react/src/util/Constants.ts
@@ -1,0 +1,1 @@
+export const BACKEND_URL = process.env.BACKEND_URL || "http://localhost:5000";
